Add files via upload
The weight of a newborn baby is one of the indicators of its health. The weight of
the newborn baby depends on numerous factors such as the mother’s health, genetics,
ethnicity, weight etc. Even more than this, another significant factor that can impact the
weight of a newborn baby is the smoking habits of the mother. Smoking is considered
harmful to health and pregnant women are often warned against it, to ensure the proper
health of their babies. In this project, we aim to determine whether smoking has any
statistically significant effect on the weight of the newborn.
Data containing the weight of the newborns along with the smoking habits of their
mothers are analyzed to see if the average weight of the babies differs significantly across
the groups corresponding to the different smoking habits of the mothers. For this, we
perform the ANOVA test on the mean weight across different groups based on smoking
habits, followed by individual pairwise tests between the different groups. Performing
pairwise tests individually increases the chance of a type 1 error for the global test. To
deal with this, we further use Bonferroni’s Correction and Tukey’s Honest Significant
Difference (HSD), to adjust the tests accordingly.
Problem statement, datasets, descriptive analysis of data, inferential statistics statistical
tests performed along with their assumptions and the results are described in the
following sections.